Login Statusabfrage implementiert

+++ app/src/main/java/com/htw/finanzplanung/Data_Access.java
@@ -27,11 +27,13 @@ public class Data_Access extends SQLiteOpenHelper{
 
     private static final String DATABASE_NAME = "Finanzplanung.sqlite";
 
-
     public Data_Access(Context context) {
         super(context, DATABASE_NAME, null, 1);
         //SQLiteDatabase db = this.getWritableDatabase();
     }
+
+
+
 
     @Override
     public void onCreate(SQLiteDatabase db) {
@@ -41,26 +43,17 @@ public class Data_Access extends SQLiteOpenHelper{
                             "  _id INTEGER PRIMARY KEY AUTOINCREMENT," +
                             "  name TEXT," +
                             "  email TEXT UNIQUE," +
-                            "  passwort TEXT NOT NULL" +
-                            "  activationcode INTEGER DEFAULT 0" +
+                            "  passwort TEXT NOT NULL," +
+                            "  loginstatus INTEGER DEFAULT 0" +
                         ")"
         );
 
-        db.execSQL("CREATE TABLE IF NOT EXISTS loginstate " +
-                        "(" +
-                        "  _id INTEGER PRIMARY KEY AUTOINCREMENT," +
-                        "  name TEXT," +
-                        "  email TEXT UNIQUE," +
-                        "  passwort TEXT NOT NULL" +
-                        "  activationcode INTEGER DEFAULT 0" +
-                        ")"
-        );
 
         db.execSQL("CREATE TABLE IF NOT EXISTS gruppe " +
                         "(" +
                         "  _id INTEGER PRIMARY KEY AUTOINCREMENT," +
                         "  name TEXT," +
-                        "  user_id INTEGER REFERENCES user(_id) ON UPDATE CASCADE ON DELETE CASCADE," +
+                        "  user_id INTEGER REFERENCES user(_id) ON UPDATE CASCADE ON DELETE CASCADE" +
                         ")"
         );
 
@@ -69,7 +62,7 @@ public class Data_Access extends SQLiteOpenHelper{
                         "  _id INTEGER PRIMARY KEY AUTOINCREMENT," +
                         "  server TEXT DEFAULT 'http://fomenko.eu/Finanzplanung/'," +
                         "  mobile_sync INTEGER DEFAULT 0," +
-                        "  user_id INTEGER REFERENCES user(_id) ON UPDATE CASCADE ON DELETE CASCADE," +
+                        "  user_id INTEGER REFERENCES user(_id) ON UPDATE CASCADE ON DELETE CASCADE" +
                         ")"
         );
 
@@ -108,7 +101,8 @@ public class Data_Access extends SQLiteOpenHelper{
 
     //GruppenVerwalten
 
-    public ArrayList<Gruppe> getMeineGruppen(Integer user_id) {
+    public ArrayList<Gruppe> getMeineGruppen() {
+        Integer user_id = getLoginState();
         ArrayList<Gruppe> Gruppen = new ArrayList<>();
         SQLiteDatabase db = this.getWritableDatabase();
 
@@ -127,15 +121,22 @@ public class Data_Access extends SQLiteOpenHelper{
         return Gruppen;
     }
 
-    public int deleteGruppe(Integer user_id, Integer gruppen_id) {
+    public int deleteGruppe(Integer gruppen_id) {
+        Integer user_id = getLoginState();
         SQLiteDatabase db = this.getWritableDatabase();
-        db.execSQL("DELETE FROM gruppe WHERE _id = " + gruppen_id + ");");
+        int zwisch = 0;
+        if(user_id.equals(getGruppenMasterID(gruppen_id))){
+            db.execSQL("DELETE FROM gruppe WHERE _id = " + gruppen_id + ");");
+            zwisch = 1;
+        }
+
         db.close();
 
-        return 0;
+        return zwisch;
     }
 
-    public int addGruppe(Integer user_id,String gruppenname){
+    public int addGruppe(String gruppenname){
+        Integer user_id = getLoginState();
         SQLiteDatabase db = this.getWritableDatabase();
 
         db.execSQL("INSERT INTO gruppe (name, user_id) VALUES ('" + gruppenname + "' , " + user_id + ");");
@@ -163,7 +164,8 @@ public class Data_Access extends SQLiteOpenHelper{
 
 
     //Finanzen
-    public int addGeldausgabe(String datum,String was, Float betrag, Integer user_id, Integer gruppen_id){
+    public int addGeldausgabe(String datum,String was, Float betrag, Integer gruppen_id){
+        Integer user_id = getLoginState();
         SQLiteDatabase db = this.getWritableDatabase();
 
         //datum TEXT as strings ("YYYY-MM-DD").
@@ -186,7 +188,7 @@ public class Data_Access extends SQLiteOpenHelper{
                 "SELECT sum(geldbetrag) AS Summe " +
                 "FROM ausgabe " +
                 "WHERE gruppe_id = " + gruppen_id +" " +
-                        "AND ausgabe.datum BETWEEN '" + startdatum + "' AND '" + enddatum + "' ", null
+                        "AND Date(ausgabe.datum) BETWEEN Date('" + startdatum + "') AND Date('" + enddatum + "') ", null
         );
         Float gesamtgeldbetrag = 0f;
         if(c.moveToFirst()){
@@ -208,7 +210,7 @@ public class Data_Access extends SQLiteOpenHelper{
                         "ON ausgabe.gruppe_id = user_ist_mitglied_in_gruppe.gruppe_id  " +
                         "WHERE user_ist_mitglied_in_gruppe.gruppe_id = " + gruppen_id + " " +
                         "AND user_ist_mitglied_in_gruppe.user_id = " + user_id + " " +
-                        "AND ausgabe.datum BETWEEN '" + startdatum + "' AND '" + enddatum + "' ", null
+                        "AND Date(ausgabe.datum) BETWEEN Date('" + startdatum + "') AND Date('" + enddatum + "') ", null
         );
         Float gesamtgeldbetrag = 0f;
         if(c.moveToFirst()){
@@ -220,7 +222,7 @@ public class Data_Access extends SQLiteOpenHelper{
         return gesamtgeldbetrag;
     }
 
-    public ArrayList<Geldausgabe> getUserGeldausgaben(Integer gruppen_id, Integer user_id){
+    public ArrayList<Geldausgabe> getUserGeldausgaben(String startdatum, String enddatum, Integer gruppen_id, Integer user_id){
         ArrayList<Geldausgabe> Ausgaben = new ArrayList<>();
         SQLiteDatabase db = this.getWritableDatabase();
 
@@ -230,7 +232,8 @@ public class Data_Access extends SQLiteOpenHelper{
                 "INNER JOIN user_ist_mitglied_in_gruppe " +
                         "ON ausgabe.gruppe_id = user_ist_mitglied_in_gruppe.gruppe_id  " +
                 "WHERE user_ist_mitglied_in_gruppe.gruppe_id = " + gruppen_id + " " +
-                        "AND user_ist_mitglied_in_gruppe.user_id = " + user_id + " ", null
+                        "AND user_ist_mitglied_in_gruppe.user_id = " + user_id + " " +
+                        "AND Date(ausgabe.datum) BETWEEN Date('" + startdatum + "') AND Date('" + enddatum + "') ", null
         );
 
         if(c.moveToFirst()){
@@ -268,7 +271,8 @@ public class Data_Access extends SQLiteOpenHelper{
     }
 
     //gibt 0 zurück wenn alles gut lief, und gibt -1 zurück wenn der user nicht der eigentümer der Gruppe ist
-    public int deleteMitglied(Integer user_id, Integer gruppen_id, Integer mitglied_id){
+    public int deleteMitglied(Integer gruppen_id, Integer mitglied_id){
+        Integer user_id = getLoginState();
         SQLiteDatabase db = this.getWritableDatabase();
         Integer zwisch = -1;
         if(user_id.intValue() == getGruppenMasterID(gruppen_id).intValue()){
@@ -280,7 +284,8 @@ public class Data_Access extends SQLiteOpenHelper{
         return zwisch;
     }
 
-    public int addGruppenMitglied(Integer user_id, Integer gruppen_id, Integer mitglied_id){
+    public int addGruppenMitglied(Integer gruppen_id, Integer mitglied_id){
+        Integer user_id = getLoginState();
         SQLiteDatabase db = this.getWritableDatabase();
 
         //datum TEXT as strings ("YYYY-MM-DD").
@@ -294,7 +299,8 @@ public class Data_Access extends SQLiteOpenHelper{
     }
 
     //Home
-    public ArrayList<Gruppe> getGruppen(Integer user_id){
+    public ArrayList<Gruppe> getGruppen(){
+        Integer user_id = getLoginState();
         ArrayList<Gruppe> Gruppen = new ArrayList<>();
         SQLiteDatabase db = this.getWritableDatabase();
 
@@ -319,7 +325,8 @@ public class Data_Access extends SQLiteOpenHelper{
         return Gruppen;
     }
 
-    public void verlasseGruppe(Integer user_id, Integer gruppen_id){
+    public void verlasseGruppe(Integer gruppen_id){
+        Integer user_id = getLoginState();
 
         SQLiteDatabase db = this.getWritableDatabase();
 
@@ -331,27 +338,54 @@ public class Data_Access extends SQLiteOpenHelper{
     //Startseite
     public Mitglied Login(String email,String passwort){
 
+
         return null;
     }
 
-    public int sendPasswortToEmail(String email){
+    public Integer getLoginState(){
+        Integer id = null;
+        SQLiteDatabase db = this.getWritableDatabase();
 
 
-        return 0;
+        Cursor c = db.rawQuery(
+                "SELECT _id " +
+                "FROM user " +
+                "WHERE loginstatus = 1 ", null
+        );
+
+
+        if(c.moveToFirst()){
+            do{
+                id = c.getInt(0);
+            }while(c.moveToNext());
+        }
+        c.close();
+        db.close();
+
+
+
+        return id;
+    }
+
+    public Boolean Logout(){
+        SQLiteDatabase db = this.getWritableDatabase();
+
+        db.execSQL(
+                "UPDATE user " +
+                        "SET loginstatus = 0 " +
+                        "WHERE loginstatus = 1 ", null
+        );
+
+        db.close();
+
+        return null;
     }
 
 
+    public String sendPasswortToEmail(String email){
+        String jsonStr;
 
-
-    public String stripHtml(String html)
-    {
-
-        return html.substring(html.indexOf("{"), html.lastIndexOf("}") + 1);
-    }
-
-    //Registration
-    public String registration(String name, String email, String passwort){
-        String url = "http://home.htw-berlin.de/~s0539589/Finanzplanung/registration.php";
+        String url = "http://home.htw-berlin.de/~s0539589/Finanzplanung/sendpassword.php";
 
         ServiceHandler sh = new ServiceHandler();
 
@@ -359,19 +393,59 @@ public class Data_Access extends SQLiteOpenHelper{
         List<NameValuePair> PHPanfrage = new ArrayList<>();
 
         PHPanfrage.add(new BasicNameValuePair("email", email));
-        PHPanfrage.add(new BasicNameValuePair("name", name));
-        PHPanfrage.add(new BasicNameValuePair("password", passwort));
 
-        String jsonStr = sh.makeServiceCall(url, ServiceHandler.POST,PHPanfrage);
-
-        Log.d("Response: ", "> " + jsonStr);
-
+        jsonStr = sh.makeServiceCall(url, ServiceHandler.POST,PHPanfrage);
 
         if (jsonStr != null) {
             try {
                 JSONObject jsonObj = new JSONObject(stripHtml(jsonStr));
-
                 jsonStr = jsonObj.getString("exception");
+            } catch (JSONException e) {
+                jsonStr = "ERROR: "+stripHtml(jsonStr);
+                e.printStackTrace();
+            }
+        } else {
+            jsonStr = "ERROR: NO INTERNET CONNECTION";
+        }
+
+        sh.destroy();
+
+        return jsonStr;
+    }
+
+
+
+
+    public String stripHtml(String html)
+    {
+        return html.substring(html.indexOf("{"), html.lastIndexOf("}") + 1);
+    }
+
+    //Registration
+    public String registration(String name, String email, String passwort, String passwortValidation){
+        String jsonStr;
+        if(passwort.equals(passwortValidation)){
+            String url = "http://home.htw-berlin.de/~s0539589/Finanzplanung/registration.php";
+
+            ServiceHandler sh = new ServiceHandler();
+
+            // Making a request to url and getting response
+            List<NameValuePair> PHPanfrage = new ArrayList<>();
+
+            PHPanfrage.add(new BasicNameValuePair("email", email));
+            PHPanfrage.add(new BasicNameValuePair("name", name));
+            PHPanfrage.add(new BasicNameValuePair("password", passwort));
+
+            jsonStr = sh.makeServiceCall(url, ServiceHandler.POST,PHPanfrage);
+
+            Log.d("Response: ", "> " + jsonStr);
+
+
+            if (jsonStr != null) {
+                try {
+                    JSONObject jsonObj = new JSONObject(stripHtml(jsonStr));
+
+                    jsonStr = jsonObj.getString("exception");
 
 /*
                 // Getting JSON Array node
@@ -398,35 +472,42 @@ public class Data_Access extends SQLiteOpenHelper{
                     contactList.add(contact);
           */
                 } catch (JSONException e) {
-                jsonStr = "ERROR: "+stripHtml(jsonStr);
-                e.printStackTrace();
+                    jsonStr = "ERROR: "+stripHtml(jsonStr);
+                    e.printStackTrace();
+                }
+            } else {
+                Log.e("ServiceHandler", "Couldn't get any data from the url");
+                jsonStr = "ERROR: NO INTERNET CONNECTION";
+
             }
-        } else {
-            Log.e("ServiceHandler", "Couldn't get any data from the url");
-            jsonStr = "ERROR: NO INTERNET CONNECTION";
+            sh.destroy();
+        }else{
+            jsonStr = "ERROR: Passwort stimmt nicht überein";
         }
 
         return jsonStr;
     }
-    public int addUser(String name, String email, String passwort){
+    public int addUser(Integer user_id, String name, String email, String passwort){
         SQLiteDatabase db = this.getWritableDatabase();
 
         //datum TEXT as strings ("YYYY-MM-DD").
-        db.execSQL("INSERT INTO user (name, email, passwort) " +
+        db.execSQL("INSERT INTO user (_id, name, email, passwort) " +
                 "VALUES (" +
+                "  " +user_id         +" , " +
                 " '" +name            +"', " +
                 " '" +email           +"', " +
-                " '" +passwort        +"' , " +
+                " '" +passwort        +"'  " +
                 ");");
         db.close();
         return 0;
     }
 
     //Settings
-    public int setNewPasswort(String newpasswort, String oldpasswort, Integer user_id){
+    public int setNewPasswort(String newpasswort, String oldpasswort){
+        Integer user_id = getLoginState();
         SQLiteDatabase db = this.getWritableDatabase();
 
-        if(validation(user_id,oldpasswort)){
+        if(validation(oldpasswort)){
             db.execSQL("UPDATE user SET password = '"+ newpasswort +"' WHERE _id = "+ user_id +";");
         }
 
@@ -434,7 +515,8 @@ public class Data_Access extends SQLiteOpenHelper{
         return 0;
 
     }
-    public Boolean validation(Integer user_id, String passwort){
+    public Boolean validation(String passwort){
+        Integer user_id = getLoginState();
         SQLiteDatabase db = this.getWritableDatabase();
 
         Cursor c = db.rawQuery(
@@ -455,10 +537,11 @@ public class Data_Access extends SQLiteOpenHelper{
         }
     }
 
-    public int setNewName(String newName, String passwort, Integer user_id){
+    public int setNewName(String newName, String passwort){
+        Integer user_id = getLoginState();
         SQLiteDatabase db = this.getWritableDatabase();
 
-        if(validation(user_id,passwort)){
+        if(validation(passwort)){
             db.execSQL("UPDATE user SET name = '"+ newName +"' WHERE _id = "+ user_id +";");
         }
 
@@ -466,10 +549,11 @@ public class Data_Access extends SQLiteOpenHelper{
         return 0;
 
     }
-    public int setNewServer(String newServer, String passwort, Integer user_id){
+    public int setNewServer(String newServer, String passwort){
+        Integer user_id = getLoginState();
         SQLiteDatabase db = this.getWritableDatabase();
 
-        if(validation(user_id,passwort)){
+        if(validation(passwort)){
             db.execSQL("UPDATE setting SET server = '"+ newServer +"' WHERE user_id = "+ user_id +";");
         }
 
@@ -477,16 +561,17 @@ public class Data_Access extends SQLiteOpenHelper{
         return 0;
 
     }
-    public void setMobileSync(Boolean mobileSync, String passwort, Integer user_id){
+    public void setMobileSync(Boolean mobileSync, String passwort){
+        Integer user_id = getLoginState();
         SQLiteDatabase db = this.getWritableDatabase();
 
-        if(validation(user_id,passwort)){
+        if(validation(passwort)){
             db.execSQL("UPDATE setting SET mobile_sync = "+ (mobileSync?1:0) +" WHERE user_id = "+ user_id +";");
         }
         db.close();
     }
-    public Boolean getMobileSyncStatus(Integer user_id){
-
+    public Boolean getMobileSyncStatus(){
+        Integer user_id = getLoginState();
         SQLiteDatabase db = this.getWritableDatabase();
         Boolean zwisch = false;
         Cursor c = db.rawQuery(
